replace graphql to github
