+++ connectors/sources/github.py
@@ -5,8 +5,6 @@
 #
 """GitHub source module responsible to fetch documents from GitHub Cloud and Server."""
 import asyncio
-import json
-import os
 import time
 from enum import Enum
 from functools import cached_property, partial
@@ -15,7 +13,7 @@ import aiofiles
 import aiohttp
 from aiofiles.os import remove
 from aiofiles.tempfile import NamedTemporaryFile
-from gidgethub import BadRequest
+from aiohttp.client_exceptions import ClientResponseError
 from gidgethub.aiohttp import GitHubAPI
 
 from connectors.logger import logger
@@ -27,10 +25,15 @@ from connectors.utils import (
     convert_to_b64,
     decode_base64_value,
     retryable,
+    ssl_context,
 )
 
 WILDCARD = "*"
 BLOB = "blob"
+GITHUB_CLOUD = "github_cloud"
+GITHUB_SERVER = "github_server"
+PULL_REQUEST_OBJECT = "pullRequest"
+REPOSITORY_OBJECT = "repository"
 
 RETRIES = 3
 RETRY_INTERVAL = 2
@@ -38,74 +41,6 @@ FILE_SIZE_LIMIT = 10485760  # ~ 10 Megabytes
 PAGE_SIZE = 100
 FORBIDDEN = 403
 
-RUNNING_FTEST = (
-    "RUNNING_FTEST" in os.environ
-)  # Flag to check if a connector is run for ftest or not.
-
-
-class ObjectType(Enum):
-    REPOSITORY = "Repository"
-    ISSUE = "Issue"
-    PR = "Pull request"
-
-
-BASE_URL = "http://127.0.0.1:9091" if RUNNING_FTEST else "https://api.github.com"
-USER_API = f"{BASE_URL}/user"
-
-ENDPOINTS = {
-    "ALL_REPOS": "/user/repos?type=all&per_page={page_size}",
-    "REPO": "repos/{repo_name}",
-    "USER": "/user",
-    "PULLS": "/repos/{repo_name}/pulls?state=all&per_page={page_size}",
-    "ISSUES": "/repos/{repo_name}/issues?state=all&per_page={page_size}",
-    "TREE": "/repos/{repo_name}/git/trees/{default_branch}?recursive=1",
-    "COMMITS": "repos/{repo_name}/commits?path={path}",
-    "COMMENTS": "/repos/{repo_name}/issues/{number}/comments?per_page={page_size}",
-    "REVIEW_COMMENTS": "/repos/{repo_name}/pulls/{number}/comments?per_page={page_size}",
-    "REVIEWS": "/repos/{repo_name}/pulls/{number}/reviews?per_page={page_size}",
-}
-
-REPO_SCHEMA = {
-    "_id": "id",
-    "_timestamp": "updated_at",
-    "name": "name",
-    "full_name": "full_name",
-    "html_url": "html_url",
-    "description": "description",
-    "visibility": "visibility",
-    "language": "language",
-    "default_branch": "default_branch",
-    "fork": "fork",
-    "open_issues": "open_issues",
-    "forks_count": "forks_count",
-    "watchers": "watchers",
-    "stargazers_count": "stargazers_count",
-    "created_at": "created_at",
-}
-ISSUE_SCHEMA = {
-    "_id": "id",
-    "_timestamp": "updated_at",
-    "number": "number",
-    "html_url": "html_url",
-    "created_at": "created_at",
-    "closed_at": "closed_at",
-    "title": "title",
-    "body": "body",
-    "state": "state",
-}
-PULL_REQUEST_SCHEMA = {
-    "_id": "id",
-    "_timestamp": "updated_at",
-    "number": "number",
-    "html_url": "html_url",
-    "created_at": "created_at",
-    "closed_at": "closed_at",
-    "title": "title",
-    "body": "body",
-    "state": "state",
-    "merged_at": "merged_at",
-    "merge_commit_sha": "merge_commit_sha",
-}
 FILE_SCHEMA = {
     "name": "name",
     "size": "size",
@@ -117,22 +52,365 @@ FILE_SCHEMA = {
 }
 
 
+class GithubQuery(Enum):
+    USER_QUERY = """
+        query {
+        viewer {
+            login
+        }
+    }
+    """
+    REPOS_QUERY = """
+    query ($login: String!, $cursor: String) {
+    user(login: $login) {
+        repositories(first: 100, after: $cursor) {
+        pageInfo {
+            hasNextPage
+            endCursor
+        }
+        nodes {
+            id
+            updatedAt
+            name
+            nameWithOwner
+            url
+            description
+            visibility
+            primaryLanguage {
+            name
+            }
+            defaultBranchRef {
+            name
+            }
+            isFork
+            stargazerCount
+            watchers {
+            totalCount
+            }
+            forkCount
+            createdAt
+            isArchived
+        }
+        }
+    }
+    }
+    """
+    REPO_QUERY = """
+    query ($owner: String!, $repositoryName: String!) {
+    repository(owner: $owner, name: $repositoryName) {
+        id
+        updatedAt
+        name
+        nameWithOwner
+        url
+        description
+        visibility
+        primaryLanguage {
+        name
+        }
+        defaultBranchRef {
+        name
+        }
+        isFork
+        stargazerCount
+        watchers {
+        totalCount
+        }
+        forkCount
+        createdAt
+        isArchived
+    }
+    }
+    """
+    PULL_REQUEST_QUERY = """
+    query ($owner: String!, $name: String!, $cursor: String) {
+    repository(owner: $owner, name: $name) {
+        pullRequests(first: 100, after: $cursor) {
+        pageInfo {
+            hasNextPage
+            endCursor
+        }
+        nodes {
+            id
+            updatedAt
+            number
+            url
+            createdAt
+            closedAt
+            title
+            body
+            state
+            mergedAt
+            assignees(first: 100) {
+            pageInfo {
+                hasNextPage
+                endCursor
+            }
+            nodes {
+                login
+            }
+            }
+            labels(first: 100) {
+            pageInfo {
+                hasNextPage
+                endCursor
+            }
+            nodes {
+                name
+                description
+            }
+            }
+            reviewRequests(first: 100) {
+            pageInfo {
+                hasNextPage
+                endCursor
+            }
+            nodes {
+                requestedReviewer {
+                ... on User {
+                    login
+                }
+                }
+            }
+            }
+            comments(first: 100) {
+            pageInfo {
+                hasNextPage
+                endCursor
+            }
+            nodes {
+                author {
+                login
+                }
+                body
+            }
+            }
+            reviews(first: 45) {
+            pageInfo {
+                hasNextPage
+                endCursor
+            }
+            nodes {
+                id
+                author {
+                login
+                }
+                state
+                body
+                comments(first: 100) {
+                pageInfo {
+                    hasNextPage
+                    endCursor
+                }
+                nodes {
+                    body
+                }
+                }
+            }
+            }
+        }
+        }
+    }
+    }
+    """
+    ISSUE_QUERY = """
+    query ($owner: String!, $name: String!, $cursor: String) {
+    repository(owner: $owner, name: $name) {
+        issues(first: 100, after: $cursor) {
+        pageInfo {
+            hasNextPage
+            endCursor
+        }
+        nodes {
+            id
+            updatedAt
+            number
+            url
+            createdAt
+            closedAt
+            title
+            body
+            state
+            assignees(first: 100) {
+            pageInfo {
+                hasNextPage
+                endCursor
+            }
+            nodes {
+                login
+            }
+            }
+            labels(first: 100) {
+            pageInfo {
+                hasNextPage
+                endCursor
+            }
+            nodes {
+                name
+                description
+            }
+            }
+            comments(first: 100) {
+            pageInfo {
+                hasNextPage
+                endCursor
+            }
+            nodes {
+                author {
+                login
+                }
+                body
+            }
+            }
+        }
+        }
+    }
+    }
+    """
+    COMMENT_QUERY = """
+    query ($owner: String!, $name: String!, $number: Int!, $cursor: String) {{
+    repository(owner: $owner, name: $name) {{
+        {object_type}(number: $number) {{
+        comments(first: 100, after: $cursor) {{
+            pageInfo {{
+            hasNextPage
+            endCursor
+            }}
+            nodes {{
+            author {{
+                login
+            }}
+            body
+            }}
+        }}
+        }}
+    }}
+    }}
+    """
+    REVIEW_QUERY = """
+    query ($owner: String!, $name: String!, $number: Int!, $cursor: String) {
+    repository(owner: $owner, name: $name) {
+        pullRequest(number: $number) {
+        reviews(first: 100, after: $cursor) {
+            pageInfo {
+            hasNextPage
+            endCursor
+            }
+            nodes {
+            id
+            author {
+                login
+            }
+            state
+            body
+            comments(first: 100) {
+                pageInfo {
+                hasNextPage
+                endCursor
+                }
+                nodes {
+                body
+                }
+            }
+            }
+        }
+        }
+    }
+    }
+    """
+    REVIEWERS_QUERY = """
+    query ($owner: String!, $name: String!, $number: Int!, $cursor: String) {
+    repository(owner: $owner, name: $name) {
+        pullRequest(number: $number) {
+        reviewRequests(first: 100, after: $cursor) {
+            pageInfo {
+            hasNextPage
+            endCursor
+            }
+            nodes {
+                requestedReviewer {
+                    ... on User {
+                    login
+                }
+            }
+            }
+        }
+        }
+    }
+    }
+    """
+    LABELS_QUERY = """
+    query ($owner: String!, $name: String!, $number: Int!, $cursor: String) {{
+    repository(owner: $owner, name: $name) {{
+        {object_type}(number: $number) {{
+        labels(first: 100, after: $cursor) {{
+            pageInfo {{
+            hasNextPage
+            endCursor
+            }}
+            nodes {{
+            name
+            }}
+        }}
+        }}
+    }}
+    }}
+    """
+    ASSIGNEES_QUERY = """
+    query ($owner: String!, $name: String!, $number: Int!, $cursor: String) {{
+    repository(owner: $owner, name: $name) {{
+        {object_type}(number: $number) {{
+        assignees(first: 100, after: $cursor) {{
+            pageInfo {{
+            hasNextPage
+            endCursor
+            }}
+            nodes {{
+            login
+            }}
+        }}
+        }}
+    }}
+    }}
+    """
+
+
+class ObjectType(Enum):
+    REPOSITORY = "Repository"
+    ISSUE = "Issue"
+    PULL_REQUEST = "Pull request"
+
+
 class UnauthorizedException(Exception):
     pass
 
 
 class GitHubClient:
     def __init__(self, configuration):
-        self.session = None
         self._sleeps = CancellableSleeps()
         self.configuration = configuration
         self._logger = logger
+        self.is_cloud = configuration["data_source"] == GITHUB_CLOUD
+        self.ssl_enabled = self.configuration["ssl_enabled"]
+        self.certificate = self.configuration["ssl_ca"]
+        self.github_url = f"{configuration['host'].rstrip('/')}/api"
         self.repos = self.configuration["repositories"]
-        self.github_token = self.configuration["github_token"]
-        self.headers = {
-            "Authorization": f"Bearer {self.github_token}",
-            "Accept": "application/vnd.github.raw",
+        self.github_token = self.configuration["token"]
+        if self.ssl_enabled and self.certificate:
+            self.ssl_ctx = ssl_context(certificate=self.certificate)
+        else:
+            self.ssl_ctx = False
+        self.endpoints = {
+            "TREE": "/api/v3/repos/{repo_name}/git/trees/{default_branch}?recursive=1",
+            "COMMITS": "api/v3/repos/{repo_name}/commits?path={path}",
         }
+        if self.is_cloud:
+            self.github_url = "https://api.github.com"
+            self.endpoints = {
+                "TREE": "/repos/{repo_name}/git/trees/{default_branch}?recursive=1",
+                "COMMITS": "/repos/{repo_name}/commits?path={path}",
+            }
+        self.user = None
 
     def set_logger(self, logger_):
         self._logger = logger_
@@ -140,27 +418,100 @@ class GitHubClient:
     def get_rate_limit_encountered(self, status_code, message):
         return status_code == FORBIDDEN and "rate limit" in str(message).lower()
 
-    async def _get_retry_after(self):
+    async def _get_retry_after(self, type):
+        current_time = time.time()
         response = await self._get_client.getitem("/rate_limit")
-        reset = response["resources"]["core"]["reset"]
+        reset = response.get("resources", {}).get(type, {}).get("reset", current_time)
         # Adding a 5 second delay to account for server delays
-        return (reset - time.time()) + 5
+        return (reset - current_time) + 5
 
-    async def _put_to_sleep(self):
-        retry_after = await self._get_retry_after()
+    async def _put_to_sleep(self, type):
+        retry_after = await self._get_retry_after(type=type)
         self._logger.debug(
             f"Connector will attempt to retry after {retry_after} seconds."
         )
         await self._sleeps.sleep(retry_after)
         raise Exception("Rate limit exceeded.")
 
+    @cached_property
+    def _get_session(self):
+        headers = {
+            "Authorization": f"Bearer {self.github_token}",
+            "Accept": "application/vnd.github.raw",
+        }
+        connector = aiohttp.TCPConnector(ssl=self.ssl_ctx)
+        timeout = aiohttp.ClientTimeout(total=None)
+        return aiohttp.ClientSession(
+            headers=headers,
+            timeout=timeout,
+            raise_for_status=True,
+            connector=connector,
+        )
+
+    @cached_property
+    def _get_client(self):
+        return GitHubAPI(
+            session=self._get_session,
+            requester="",
+            oauth_token=self.github_token,
+            base_url=self.github_url,
+        )
+
     @retryable(
         retries=RETRIES,
         interval=RETRY_INTERVAL,
         strategy=RetryStrategy.EXPONENTIAL_BACKOFF,
     )
-    async def get_response(self, resource):
-        """Execute request using getitem method of GitHubAPI.
+    async def post(self, query_data, need_headers=False):
+        """Invoke GraphQL request to fetch repositories, pull requests, and issues.
+
+        Args:
+            query_data: Dictionary comprising of query and variables for the GraphQL request.
+
+        Raises:
+            UnauthorizedException: Unauthorized exception
+            exception: An instance of an exception class.
+
+        Yields:
+            dictionary: Client response
+        """
+        url = f"{self.github_url}/graphql"
+        try:
+            async with self._get_session.post(
+                url=url, json=query_data, ssl=self.ssl_ctx
+            ) as response:
+                json_response = await response.json()
+                if not json_response.get("errors"):
+                    return (
+                        (json_response, response.headers)
+                        if need_headers
+                        else json_response
+                    )
+                for error in json_response.get("errors"):
+                    if (
+                        error.get("type") == "RATE_LIMITED"
+                        and "api rate limit exceeded" in error.get("message").lower()
+                    ):
+                        await self._put_to_sleep(type="graphql")
+                raise Exception(
+                    f"Error while executing query. Exception: {json_response['errors']}"
+                )
+        except ClientResponseError as exception:
+            if exception.status == 401:
+                raise UnauthorizedException(
+                    "Your Github token is either expired or revoked. Please check again."
+                ) from exception
+        except Exception:
+            raise
+
+    @retryable(
+        retries=RETRIES,
+        interval=RETRY_INTERVAL,
+        strategy=RetryStrategy.EXPONENTIAL_BACKOFF,
+    )
+    async def get_github_item(self, resource):
+        """Execute request using getitem method of GitHubAPI which is using REST API.
+        Using Rest API for fetching files and folder along with content.
 
         Args:
             resource (str): API to get the response
@@ -170,104 +521,105 @@ class GitHubClient:
         """
         try:
             return await self._get_client.getitem(url=resource)
-        except BadRequest as exception:
-            if self.get_rate_limit_encountered(exception.status_code, exception):
-                await self._put_to_sleep()
-            elif self.is_unauthorized(exception=exception):
+        except ClientResponseError as exception:
+            if exception.status == 401:
                 raise UnauthorizedException(
                     "Your Github token is either expired or revoked. Please check again."
                 ) from exception
+            elif self.get_rate_limit_encountered(exception.status, exception):
+                await self._put_to_sleep("core")
             else:
                 raise
         except Exception:
             raise
 
-    @retryable(
-        retries=RETRIES,
-        interval=RETRY_INTERVAL,
-        strategy=RetryStrategy.EXPONENTIAL_BACKOFF,
-    )
-    async def get_response_headers(self, method, url):
-        """Execute request using _request method of GitHubAPI and returns response headers.
+    def get_page_info(self, response, keys):
+        """Get page information from the response.
 
         Args:
-            method (str): Method of the request
-            url (str): API to get the response
+            response (dict): Response of GitHub
+            Keys (list): List of fields to get pageInfo
 
-        Returns:
-            dict: Response of the headers
+        Return:
+            dict: dictionary containing page information.
         """
-        status_code, response_headers, data = await self._get_client._request(
-            method=method, url=url, headers=self.headers
-        )
-        if status_code == 200:
-            return response_headers
-        try:
-            data = data.decode("utf-8")
-            message = json.loads(data)
-        except ValueError as exception:
-            self._logger.exception(
-                f"Error while loading the response data. Error: {exception}"
-            )
-            raise
-        if self.get_rate_limit_encountered(status_code, message):
-            await self._put_to_sleep()
-        else:
-            raise Exception(
-                f"Something went wrong while executing request Exception: {message['message']}"
-            )
+        current_level = response.get("data")
+        for key in keys:
+            current_level = current_level.get(key)
+            if current_level is None:
+                break
+        return current_level.get("pageInfo")
 
-    async def get_paginated_response(self, resource):
-        """Execute paginated request using getitem method of GithubAPI.
+    async def paginated_api_call(self, variables, query, keys):
+        """Make a paginated API call for fetching GitHub objects.
 
         Args:
-            resource (str): API to get the response
+            variables (dict): Variables for Graphql API
+            query (string): Graphql Query
+            keys (list): List of fields to get pageInfo
 
         Yields:
-            dict: Response of the request
+            dict: dictionary containing response of GitHub.
         """
-        page = 1
         while True:
-            url = f"{resource}&page={page}"
-            documents = await self.get_response(resource=url)
-            if not documents:
-                break
-            for document in documents:
-                yield document
-            page += 1
-
-    async def get_user_repos(self):
-        async for response in self.get_paginated_response(
-            resource=ENDPOINTS["ALL_REPOS"].format(page_size=PAGE_SIZE)
-        ):
+            query_data = {"query": query, "variables": variables}
+            response = await self.post(query_data=query_data)
             yield response
 
-    async def get_foreign_repo(self, repo_name):
-        return await self.get_response(
-            resource=ENDPOINTS["REPO"].format(repo_name=repo_name)
-        )
+            page_info = self.get_page_info(response=response, keys=keys)
+            if not page_info.get("hasNextPage"):
+                break
+            variables["cursor"] = page_info["endCursor"]
 
-    @cached_property
-    def _get_client(self):
-        self.session = aiohttp.ClientSession(headers=self.headers)
-        return GitHubAPI(
-            self.session,
-            requester="",
-            oauth_token=self.github_token,
-            base_url=BASE_URL,
+    def get_repo_details(self, repo_name):
+        return repo_name.split("/")
+
+    async def get_user_repos(self):
+        repo_variables = {
+            "login": self.user,
+            "cursor": None,
+        }
+        async for response in self.paginated_api_call(
+            variables=repo_variables,
+            query=GithubQuery.REPOS_QUERY.value,
+            keys=["user", "repositories"],
+        ):
+            for repo in (
+                response.get("data", {})  # pyright: ignore
+                .get("user", {})
+                .get("repositories", {})
+                .get("nodes")
+            ):
+                yield repo
+
+    async def get_foreign_repo(self, repo_name):
+        owner, repo = self.get_repo_details(repo_name=repo_name)
+        repo_variables = {"owner": owner, "repositoryName": repo}
+        query_data = {
+            "query": GithubQuery.REPO_QUERY.value,
+            "variables": repo_variables,
+        }
+        repo_response = await self.post(query_data=query_data)
+        return repo_response.get("data", {}).get(REPOSITORY_OBJECT)  # pyright: ignore
+
+    async def get_logged_in_user(self):
+        query_data = {
+            "query": GithubQuery.USER_QUERY.value,
+            "variables": None,
+        }
+        response = await self.post(query_data=query_data)
+        return (
+            response.get("data", {}).get("viewer", {}).get("login")  # pyright: ignore
         )
 
     async def ping(self):
-        await self.get_response(resource=ENDPOINTS["USER"])
+        query_data = {"query": GithubQuery.USER_QUERY.value, "variables": None}
+        await self.post(query_data=query_data)
 
     async def close(self):
         self._sleeps.cancel()
-        if self.session is not None:
-            await self.session.close()
-            del self._get_client
-
-    def is_unauthorized(self, exception):
-        return exception.status_code.value == 401
+        await self._get_session.close()
+        del self._get_session
 
 
 class GitHubDataSource(BaseDataSource):
@@ -298,25 +650,57 @@ class GitHubDataSource(BaseDataSource):
             dict: Default configuration.
         """
         return {
-            "github_token": {
-                "label": "GitHub Token",
-                "sensitive": True,
+            "data_source": {
+                "display": "dropdown",
+                "label": "GitHub data source",
+                "options": [
+                    {"label": "GitHub Cloud", "value": GITHUB_CLOUD},
+                    {"label": "GitHub Server", "value": GITHUB_SERVER},
+                ],
                 "order": 1,
+                "type": "str",
+                "value": GITHUB_SERVER,
+            },
+            "host": {
+                "depends_on": [{"field": "data_source", "value": GITHUB_SERVER}],
+                "label": "GitHub URL",
+                "order": 2,
+                "type": "str",
+                "value": "http://127.0.0.1:9091",
+            },
+            "token": {
+                "label": "GitHub Token",
+                "order": 3,
+                "sensitive": True,
                 "type": "str",
                 "value": "changeme",
             },
             "repositories": {
                 "display": "textarea",
                 "label": "List of repositories",
-                "order": 2,
+                "order": 4,
                 "type": "list",
                 "value": WILDCARD,
+            },
+            "ssl_enabled": {
+                "display": "toggle",
+                "label": "Enable SSL",
+                "order": 5,
+                "type": "bool",
+                "value": False,
+            },
+            "ssl_ca": {
+                "depends_on": [{"field": "ssl_enabled", "value": True}],
+                "label": "SSL certificate",
+                "order": 6,
+                "type": "str",
+                "value": "",
             },
             "retry_count": {
                 "display_value": RETRIES,
                 "display": "numeric",
                 "label": "Maximum retries per request",
-                "order": 3,
+                "order": 7,
                 "required": False,
                 "type": "int",
                 "ui_restrictions": ["advanced"],
@@ -331,27 +715,27 @@ class GitHubDataSource(BaseDataSource):
                 "Checking if there are any inaccessible repositories configured"
             )
             foreign_repos, configured_repos = [], []
-            self.user = await self.github_client.get_response(
-                resource=ENDPOINTS["USER"]
-            )
+            self.github_client.user = await self.github_client.get_logged_in_user()
 
             for repo_name in self.github_client.repos:
                 if repo_name not in ["", None]:
                     if "/" in repo_name:
                         foreign_repos.append(repo_name)
                     else:
-                        configured_repos.append(f"{self.user.get('login')}/{repo_name}")
-
+                        configured_repos.append(
+                            f"{self.github_client.user}/{repo_name}"
+                        )
             async for repo in self.github_client.get_user_repos():
-                self.user_repos[repo["full_name"]] = repo
+                self.user_repos[repo["nameWithOwner"]] = repo
             invalid_repos = list(set(configured_repos) - set(self.user_repos.keys()))
 
             for repo_name in foreign_repos:
                 try:
                     self.foreign_repos[
                         repo_name
-                    ] = await self.github_client.get_foreign_repo(repo_name)
-                except BadRequest:
+                    ] = await self.github_client.get_foreign_repo(repo_name=repo_name)
+                except Exception:
+                    self._logger.debug(f"Detected invalid repository: {repo_name}.")
                     invalid_repos.append(repo_name)
             return invalid_repos
         except Exception as exception:
@@ -366,10 +750,11 @@ class GitHubDataSource(BaseDataSource):
         Raises:
             ConfigurableFieldValueError: Insufficient privileges error.
         """
-        response_headers = await self.github_client.get_response_headers(
-            method="get", url=USER_API
+        query_data = {"query": GithubQuery.USER_QUERY.value, "variables": None}
+        _, headers = await self.github_client.post(  # pyright: ignore
+            query_data=query_data, need_headers=True
         )
-        if "repo" not in response_headers.get("X-OAuth-Scopes"):  # pyright: ignore
+        if "repo" not in headers.get("X-OAuth-Scopes"):  # pyright: ignore
             raise ConfigurableFieldValueError(
                 "Configured token does not have required rights to fetch the content"
             )
@@ -399,94 +784,97 @@ class GitHubDataSource(BaseDataSource):
             raise
 
     def adapt_gh_doc_to_es_doc(self, github_document, schema):
-        document = {}
-        for es_field, github_field in schema.items():
-            document[es_field] = github_document[github_field]
-        return document
-
-    def _comment_doc(self, comment):
         return {
-            "author": comment["user"]["login"],
-            "body": comment["body"],
+            es_field: github_document[github_field]
+            for es_field, github_field in schema.items()
         }
 
-    def _review_doc(self, review):
+    def _prepare_pull_request_doc(self, pull_request, reviews):
         return {
-            "author": review["user"]["login"],
-            "body": review["body"],
-            "state": review["state"],
+            "_id": pull_request.pop("id"),
+            "_timestamp": pull_request.pop("updatedAt"),
+            "type": ObjectType.PULL_REQUEST.value,
+            "issue_comments": pull_request.get("comments", {}).get("nodes"),
+            "reviews_comments": reviews,
+            "labels_field": pull_request.get("labels", {}).get("nodes"),
+            "assignees_list": pull_request.get("assignees", {}).get("nodes"),
+            "requested_reviewers": pull_request.get("reviewRequests", {}).get("nodes"),
         }
 
-    def _pull_request_doc(self, pull, repo_name, review_comments, comments, reviews):
+    def _prepare_issue_doc(self, issue):
         return {
-            "type": ObjectType.PR.value,
-            "owner": pull["user"]["login"],
-            "repository": repo_name,
-            "head_label": pull["head"]["label"],
-            "base_label": pull["base"]["label"],
-            "assignees": [assignee["login"] for assignee in pull["assignees"]],
-            "requested_reviewers": [
-                reviewer["login"] for reviewer in pull["requested_reviewers"]
-            ],
-            "requested_teams": [team["name"] for team in pull["requested_teams"]],
-            "labels": [label["name"] for label in pull["labels"]],
-            "review_comments": review_comments,
-            "comments": comments,
-            "reviews": reviews,
-        }
-
-    def _issue_doc(self, issue, repo_name, comments):
-        return {
+            "_id": issue.pop("id"),
             "type": ObjectType.ISSUE.value,
-            "owner": issue["user"]["login"],
-            "repository": repo_name,
-            "assignees": [assignee["login"] for assignee in issue["assignees"]],
-            "labels": [label["name"] for label in issue["labels"]],
-            "comments": comments,
+            "_timestamp": issue.pop("updatedAt"),
+            "issue_comments": issue.get("comments", {}).get("nodes"),
+            "labels_field": issue.get("labels", {}).get("nodes"),
+            "assignees_list": issue.get("assignees", {}).get("nodes"),
         }
 
-    async def fetch_repos(self):
+    def _prepare_review_doc(self, review):
+        return {
+            "author": review.get("author").get("login"),
+            "body": review.get("body"),
+            "state": review.get("state"),
+            "comments": review.get("comments").get("nodes"),
+        }
+
+    async def _get_personal_repos(self):
+        if not self.user_repos:
+            async for repo_object in self.github_client.get_user_repos():
+                self.user_repos[repo_object["nameWithOwner"]] = repo_object
+        for repo_object in self.user_repos.values():
+            repo_object.update(
+                {
+                    "_id": repo_object.pop("id"),
+                    "_timestamp": repo_object.pop("updatedAt"),
+                    "type": ObjectType.REPOSITORY.value,
+                }
+            )
+            yield repo_object
+
+    async def _get_configured_repos(self):
+        for repo_name in self.github_client.repos:
+            if repo_name in ["", None]:
+                continue
+
+            # Converting the local repository names to username/repo_name format.
+            if "/" not in repo_name:
+                repo_name = f"{self.github_client.user}/{repo_name}"
+            repo_object = self.foreign_repos.get(repo_name) or self.user_repos.get(
+                repo_name
+            )
+
+            if not repo_object:
+                owner, repo = self.github_client.get_repo_details(repo_name=repo_name)
+                query_data = {
+                    "query": GithubQuery.REPO_QUERY.value,
+                    "variables": {"owner": owner, "repositoryName": repo},
+                }
+                response = await self.github_client.post(query_data=query_data)
+                repo_object = response.get("data", {}).get(  # pyright: ignore
+                    REPOSITORY_OBJECT
+                )
+            repo_object.update(
+                {
+                    "_id": repo_object.pop("id"),
+                    "_timestamp": repo_object.pop("updatedAt"),
+                    "type": ObjectType.REPOSITORY.value,
+                }
+            )
+            yield repo_object
+
+    async def _fetch_repos(self):
         try:
+            if self.github_client.user is None:
+                self.github_client.user = await self.github_client.get_logged_in_user()
+
             if self.github_client.repos == [WILDCARD]:
-                if not self.user_repos:
-                    async for repo in self.github_client.get_user_repos():
-                        self.user_repos[repo["full_name"]] = repo
-                for repo in self.user_repos.values():
-                    document = self.adapt_gh_doc_to_es_doc(
-                        github_document=repo, schema=REPO_SCHEMA
-                    )
-                    document["owner"] = repo["owner"]["login"]
-                    document["type"] = ObjectType.REPOSITORY.value
-                    yield document
+                async for repo_object in self._get_personal_repos():
+                    yield repo_object
             else:
-                if self.user is None:
-                    self.user = await self.github_client.get_response(
-                        resource=ENDPOINTS["USER"]
-                    )
-
-                for repo_name in self.github_client.repos:
-                    if repo_name in ["", None]:
-                        continue
-
-                    # Converting the local repository names to username/repo_name format.
-                    if "/" not in repo_name:
-                        repo_name = f"{self.user.get('login')}/{repo_name}"
-
-                    repo_object = self.foreign_repos.get(
-                        repo_name
-                    ) or self.user_repos.get(repo_name)
-                    if not repo_object:
-                        repo_object = await self.github_client.get_response(
-                            resource=ENDPOINTS["REPO"].format(repo_name=repo_name)
-                        )
-
-                    document = self.adapt_gh_doc_to_es_doc(
-                        github_document=repo_object,
-                        schema=REPO_SCHEMA,
-                    )
-                    document["owner"] = repo_object["owner"]["login"]  # pyright: ignore
-                    document["type"] = ObjectType.REPOSITORY.value
-                    yield document
+                async for repo_object in self._get_configured_repos():
+                    yield repo_object
         except UnauthorizedException:
             raise
         except Exception as exception:
@@ -494,95 +882,179 @@ class GitHubDataSource(BaseDataSource):
                 f"Something went wrong while fetching the repository. Exception: {exception}"
             )
 
-    async def fetch_pull_requests(self, repo_name):
-        try:
-            async for pull in self.github_client.get_paginated_response(
-                resource=ENDPOINTS["PULLS"].format(
-                    repo_name=repo_name, page_size=PAGE_SIZE
-                )
-            ):
-                document = self.adapt_gh_doc_to_es_doc(
-                    github_document=pull, schema=PULL_REQUEST_SCHEMA
-                )
-                comments, review_comments, reviews = [], [], []
-                async for review_comment in self.github_client.get_paginated_response(
-                    ENDPOINTS["REVIEW_COMMENTS"].format(
-                        repo_name=repo_name, number=pull["number"], page_size=PAGE_SIZE
-                    )
-                ):
-                    review_comments.append(self._comment_doc(review_comment))
-                async for comment in self.github_client.get_paginated_response(
-                    ENDPOINTS["COMMENTS"].format(
-                        repo_name=repo_name, number=pull["number"], page_size=PAGE_SIZE
-                    )
-                ):
-                    comments.append(self._comment_doc(comment))
-                async for review in self.github_client.get_paginated_response(
-                    ENDPOINTS["REVIEWS"].format(
-                        repo_name=repo_name, number=pull["number"], page_size=PAGE_SIZE
-                    )
-                ):
-                    reviews.append(self._review_doc(review))
-                document.update(
-                    self._pull_request_doc(
-                        pull, repo_name, review_comments, comments, reviews
-                    )
-                )
-                yield document
-        except UnauthorizedException:
-            raise
-        except Exception as exception:
-            self._logger.warning(
-                f"Something went wrong while fetching the pull requests of {repo_name}. Exception: {exception}"
-            )
+    async def _fetch_remaining_data(
+        self, variables, object_type, query, field_type, keys
+    ):
+        async for response in self.github_client.paginated_api_call(
+            variables=variables, query=query, keys=keys
+        ):
+            yield response.get("data", {}).get(  # pyright: ignore
+                REPOSITORY_OBJECT, {}
+            ).get(object_type, {}).get(field_type, {}).get("nodes")
 
-    async def fetch_issues(self, repo_name):
-        try:
-            async for issue in self.github_client.get_paginated_response(
-                resource=ENDPOINTS["ISSUES"].format(
-                    repo_name=repo_name, page_size=PAGE_SIZE
-                )
+    async def _fetch_remaining_fields(
+        self, type_obj, object_type, owner, repo, field_type
+    ):
+        sample_dict = {
+            "reviews": {
+                "query": GithubQuery.REVIEW_QUERY.value,
+                "es_field": "reviews_comments",
+            },
+            "reviewRequests": {
+                "query": GithubQuery.REVIEWERS_QUERY.value,
+                "es_field": "requested_reviewers",
+            },
+            "comments": {
+                "query": GithubQuery.COMMENT_QUERY.value.format(
+                    object_type=object_type
+                ),
+                "es_field": "issue_comments",
+            },
+            "labels": {
+                "query": GithubQuery.LABELS_QUERY.value.format(object_type=object_type),
+                "es_field": "labels_field",
+            },
+            "assignees": {
+                "query": GithubQuery.ASSIGNEES_QUERY.value.format(
+                    object_type=object_type
+                ),
+                "es_field": "assignees_list",
+            },
+        }
+        page_info = type_obj.get(field_type, {}).get("pageInfo", {})
+        if page_info.get("hasNextPage"):
+            variables = {
+                "owner": owner,
+                "name": repo,
+                "number": type_obj.get("number"),
+                "cursor": page_info.get("endCursor"),
+            }
+            async for response in self._fetch_remaining_data(
+                variables=variables,
+                object_type=object_type,
+                query=sample_dict[field_type]["query"],
+                field_type=field_type,
+                keys=[REPOSITORY_OBJECT, object_type, field_type],
             ):
-                if issue.get("pull_request") is not None:
-                    continue
-                document = self.adapt_gh_doc_to_es_doc(
-                    github_document=issue, schema=ISSUE_SCHEMA
-                )
-                comments = []
-                async for comment in self.github_client.get_paginated_response(
-                    resource=ENDPOINTS["COMMENTS"].format(
-                        repo_name=repo_name, number=issue["number"], page_size=PAGE_SIZE
-                    )
-                ):
-                    comments.append(self._comment_doc(comment))
-                document.update(self._issue_doc(issue, repo_name, comments))
-                yield document
-        except UnauthorizedException:
-            raise
-        except Exception as exception:
-            self._logger.warning(
-                f"Something went wrong while fetching the issues of {repo_name}. Exception: {exception}"
-            )
+                if field_type == "reviews":
+                    for review in response:
+                        type_obj["reviews_comments"].append(
+                            self._prepare_review_doc(review=review)
+                        )
+                else:
+                    type_obj[sample_dict[field_type]["es_field"]].extend(response)
 
-    async def fetch_last_commit_timestamp(self, repo_name, path):
-        commits = await self.github_client.get_response(
-            resource=ENDPOINTS["COMMITS"].format(repo_name=repo_name, path=path)
+    async def _extract_pull_request(self, pull_request, owner, repo):
+        reviews = [
+            self._prepare_review_doc(review=review)
+            for review in pull_request.get("reviews", {}).get("nodes")
+        ]
+        pull_request.update(
+            self._prepare_pull_request_doc(pull_request=pull_request, reviews=reviews)
         )
+        for field in ["comments", "reviewRequests", "labels", "assignees", "reviews"]:
+            await self._fetch_remaining_fields(
+                type_obj=pull_request,
+                object_type=PULL_REQUEST_OBJECT,
+                owner=owner,
+                repo=repo,
+                field_type=field,
+            )
+            pull_request.pop(field)
+        yield pull_request
 
-        # Get last commit timstamp from the list of the commits
-        return commits[0]["commit"]["committer"]["date"]  # pyright: ignore
-
-    async def fetch_files(self, repo_name, default_branch):
+    async def _fetch_pull_requests(self, repo_name):
         try:
-            file_tree = await self.github_client.get_response(
-                resource=ENDPOINTS["TREE"].format(
+            owner, repo = self.github_client.get_repo_details(repo_name=repo_name)
+            pull_request_variables = {
+                "owner": owner,
+                "name": repo,
+                "cursor": None,
+            }
+            async for response in self.github_client.paginated_api_call(
+                variables=pull_request_variables,
+                query=GithubQuery.PULL_REQUEST_QUERY.value,
+                keys=[REPOSITORY_OBJECT, "pullRequests"],
+            ):
+                for pull_request in (
+                    response.get("data", {})  # pyright: ignore
+                    .get(REPOSITORY_OBJECT, {})
+                    .get("pullRequests", {})
+                    .get("nodes")
+                ):
+                    async for pull_request_doc in self._extract_pull_request(
+                        pull_request=pull_request, owner=owner, repo=repo
+                    ):
+                        yield pull_request_doc
+        except UnauthorizedException:
+            raise
+        except Exception as exception:
+            self._logger.warning(
+                f"Something went wrong while fetching the pull requests. Exception: {exception}"
+            )
+
+    async def _extract_issues(self, response, owner, repo):
+        for issue in (
+            response.get("data", {})
+            .get(REPOSITORY_OBJECT, {})
+            .get("issues", {})
+            .get("nodes")
+        ):
+            issue.update(self._prepare_issue_doc(issue=issue))
+            for field in ["comments", "labels", "assignees"]:
+                await self._fetch_remaining_fields(
+                    type_obj=issue,
+                    object_type="issue",
+                    owner=owner,
+                    repo=repo,
+                    field_type=field,
+                )
+                issue.pop(field)
+            yield issue
+
+    async def _fetch_issues(self, repo_name):
+        try:
+            owner, repo = self.github_client.get_repo_details(repo_name=repo_name)
+            issue_variables = {
+                "owner": owner,
+                "name": repo,
+                "cursor": None,
+            }
+            async for response in self.github_client.paginated_api_call(
+                variables=issue_variables,
+                query=GithubQuery.ISSUE_QUERY.value,
+                keys=[REPOSITORY_OBJECT, "issues"],
+            ):
+                async for issue in self._extract_issues(
+                    response=response, owner=owner, repo=repo
+                ):
+                    yield issue
+        except UnauthorizedException:
+            raise
+        except Exception as exception:
+            self._logger.warning(
+                f"Something went wrong while fetching the issues. Exception: {exception}"
+            )
+
+    async def _fetch_last_commit_timestamp(self, repo_name, path):
+        commit, *_ = await self.github_client.get_github_item(  # pyright: ignore
+            resource=self.github_client.endpoints["COMMITS"].format(
+                repo_name=repo_name, path=path
+            )
+        )
+        return commit["commit"]["committer"]["date"]
+
+    async def _fetch_files(self, repo_name, default_branch):
+        try:
+            file_tree = await self.github_client.get_github_item(
+                resource=self.github_client.endpoints["TREE"].format(
                     repo_name=repo_name, default_branch=default_branch
                 )
             )
 
-            for repo_object in file_tree.get("tree"):
-                last_commit_timestamp = await self.fetch_last_commit_timestamp(
-                    repo_name, repo_object["path"]
+            for repo_object in file_tree.get("tree", []):
+                last_commit_timestamp = await self._fetch_last_commit_timestamp(
+                    repo_name=repo_name, path=repo_object["path"]
                 )
                 if repo_object["type"] == BLOB:
                     file_name = repo_object["path"].split("/")[-1]
@@ -607,8 +1079,12 @@ class GitHubDataSource(BaseDataSource):
                     document["_id"] = f"{repo_name}/{repo_object['path']}"
                     yield document, repo_object
                 else:
-                    repo_object["_id"] = f"{repo_name}/{repo_object['path']}"
-                    repo_object["_timestamp"] = last_commit_timestamp
+                    repo_object.update(
+                        {
+                            "_id": f"{repo_name}/{repo_object['path']}",
+                            "_timestamp": last_commit_timestamp,
+                        }
+                    )
                     yield repo_object, None
         except UnauthorizedException:
             raise
@@ -616,6 +1092,51 @@ class GitHubDataSource(BaseDataSource):
             self._logger.warning(
                 f"Something went wrong while fetching the files of {repo_name}. Exception: {exception}"
             )
+
+    async def _get_document_with_content(self, url, attachment_name, document):
+        file_data = await self.github_client.get_github_item(resource=url)
+        temp_filename = ""
+        async with NamedTemporaryFile(mode="wb", delete=False) as async_buffer:
+            await async_buffer.write(
+                decode_base64_value(content=file_data["content"])  # pyright: ignore
+            )
+            temp_filename = str(async_buffer.name)
+
+        self._logger.debug(f"Calling convert_to_b64 for file : {attachment_name}")
+        await asyncio.to_thread(convert_to_b64, source=temp_filename)
+        async with aiofiles.open(file=temp_filename, mode="r") as async_buffer:
+            # base64 on macOS will add a EOL, so we strip() here
+            document["_attachment"] = (await async_buffer.read()).strip()
+
+        try:
+            await remove(temp_filename)
+        except Exception as exception:
+            self._logger.warning(
+                f"Could not remove file from: {temp_filename}. Error: {exception}"
+            )
+        return document
+
+    def _pre_checks_for_get_content(
+        self, attachment_extension, attachment_name, attachment_size
+    ):
+        if attachment_extension == "":
+            self._logger.warning(
+                f"Files without extension are not supported by TIKA, skipping {attachment_name}."
+            )
+            return
+
+        elif attachment_extension.lower() not in TIKA_SUPPORTED_FILETYPES:
+            self._logger.warning(
+                f"Files with the extension {attachment_extension} are not supported by TIKA, skipping {attachment_name}."
+            )
+            return
+
+        if attachment_size > FILE_SIZE_LIMIT:
+            self._logger.warning(
+                f"File size {attachment_size} of file {attachment_name} is larger than {FILE_SIZE_LIMIT} bytes. Discarding file content"
+            )
+            return
+        return True
 
     async def get_content(self, attachment, timestamp=None, doit=False):
         """Extracts the content for Apache TIKA supported file types.
@@ -635,22 +1156,11 @@ class GitHubDataSource(BaseDataSource):
         attachment_name = attachment["name"]
         attachment_extension = attachment["extension"]
 
-        if attachment_extension == "":
-            self._logger.warning(
-                f"Files without extension are not supported by TIKA, skipping {attachment_name}."
-            )
-            return
-
-        if attachment_extension.lower() not in TIKA_SUPPORTED_FILETYPES:
-            self._logger.warning(
-                f"Files with the extension {attachment_extension} are not supported by TIKA, skipping {attachment_name}."
-            )
-            return
-
-        if attachment_size > FILE_SIZE_LIMIT:
-            self._logger.warning(
-                f"File size {attachment_size} of file {attachment_name} is larger than {FILE_SIZE_LIMIT} bytes. Discarding file content"
-            )
+        if not self._pre_checks_for_get_content(
+            attachment_extension=attachment_extension,
+            attachment_name=attachment_name,
+            attachment_size=attachment_size,
+        ):
             return
 
         self._logger.debug(f"Downloading {attachment_name}")
@@ -659,27 +1169,12 @@ class GitHubDataSource(BaseDataSource):
             "_id": f"{attachment['repo_name']}/{attachment['name']}",
             "_timestamp": attachment["_timestamp"],
         }
-        file_data = await self.github_client.get_response(resource=attachment["url"])
-        temp_filename = ""
-        async with NamedTemporaryFile(mode="wb", delete=False) as async_buffer:
-            await async_buffer.write(
-                decode_base64_value(content=file_data["content"])  # pyright: ignore
-            )
-            temp_filename = str(async_buffer.name)
 
-        self._logger.debug(f"Calling convert_to_b64 for file : {attachment_name}")
-
-        await asyncio.to_thread(convert_to_b64, source=temp_filename)
-        async with aiofiles.open(file=temp_filename, mode="r") as async_buffer:
-            # base64 on macOS will add a EOL, so we strip() here
-            document["_attachment"] = (await async_buffer.read()).strip()
-        try:
-            await remove(temp_filename)
-        except Exception as exception:
-            self._logger.warning(
-                f"Could not remove file from: {temp_filename}. Error: {exception}"
-            )
-        return document
+        return await self._get_document_with_content(
+            url=attachment["url"],
+            attachment_name=attachment_name,
+            document=document,
+        )
 
     async def get_docs(self, filtering=None):
         """Executes the logic to fetch GitHub objects in async manner.
@@ -690,22 +1185,28 @@ class GitHubDataSource(BaseDataSource):
         Yields:
             dict: Documents from GitHub.
         """
-        async for repo in self.fetch_repos():
+        async for repo in self._fetch_repos():
             yield repo, None
-            repo_name = repo["full_name"]
+            repo_name = repo.get("nameWithOwner")
+            default_branch = (
+                repo.get("defaultBranchRef", {}).get("name")
+                if repo.get("defaultBranchRef")
+                else None
+            )
 
-            async for pull_request in self.fetch_pull_requests(repo_name=repo_name):
+            async for pull_request in self._fetch_pull_requests(repo_name=repo_name):
                 yield pull_request, None
 
-            async for issue in self.fetch_issues(repo_name=repo_name):
+            async for issue in self._fetch_issues(repo_name=repo_name):
                 yield issue, None
 
-            async for file_document, attachment_metadata in self.fetch_files(
-                repo_name=repo_name, default_branch=repo["default_branch"]
-            ):
-                if file_document["type"] == BLOB:
-                    yield file_document, partial(
-                        self.get_content, attachment=attachment_metadata
-                    )
-                else:
-                    yield file_document, None
+            if default_branch:
+                async for file_document, attachment_metadata in self._fetch_files(
+                    repo_name=repo_name, default_branch=default_branch
+                ):
+                    if file_document["type"] == BLOB:
+                        yield file_document, partial(
+                            self.get_content, attachment=attachment_metadata
+                        )
+                    else:
+                        yield file_document, None
